adding CFN support + misc fixes
git-svn-id: https://vgmstream.svn.sourceforge.net/svnroot/vgmstream@205 51a99a44-fe44-0410-b1ba-c3e57ba2b86b

+++ src/meta/ea_header.c
@@ -226,8 +226,8 @@ VGMSTREAM * init_vgmstream_ea(STREAMFILE *streamFile) {
 
 		do {
 			ea_block_update(vgmstream->next_block_offset,vgmstream);
-			vgmstream->num_samples+=vgmstream->current_block_size*28;		
-		} while(vgmstream->next_block_offset<(file_length-block_length));
+			vgmstream->num_samples+=(int32_t)vgmstream->current_block_size*28;		
+		} while(vgmstream->next_block_offset<(off_t)(file_length-block_length));
 	}
 
 	ea_block_update(start_offset+header_length,vgmstream);
